requirements/development.txt
-r base.txt  # Include all base dependencies

# Development tools
django-extensions==3.1.5
drf-yasg==1.21.11
drf-spectacular==0.29.0
ipython==8.0.0
black==23.0.0
flake8==6.0.0
pytest==9.0.2
pytest-django==4.7.0
autopep8==2.0.4

# Database & Cache
django-redis==5.2.0
redis==5.2.1

# JWT Authentication
djangorestframework_simplejwt==5.5.1
PyJWT==2.9.0

# Documentation & APIs
coreapi==2.3.3
coreschema==0.0.4
uritemplate==4.2.0

# Testing
factory-boy==3.3.0
Faker==20.0.0
coverage==7.3.2

# Payment & Email
chargily_pay==2.1.3
resend==2.19.0
requests==2.31.0

# Image Processing (optional for development)
opencv-python==4.11.0.86
pytesseract==0.3.13
numpy==2.0.2
